Added nArgs field for aliases
Aliases now require an nArgs field, and they will gobble that number of
arguments from whatever input string is given. e.g., if the doSomething
command took 3 arguments, then typing
 $ doSomething ayy bee cee dee eee eff
will split the arguments as:
 ayy bee cee dee eee eff
 ### ### ###############
  1   2         3

In attempting to divine apt names for each phase of the evaluator, I
went a bit overboard ... This may need to be dialed back; there are
probably too many functions, and their naming and renaming of variables
is unecessarily verbose at a glance.