Dynamic slurping and spewing, depending on alias
Callback functions now recieve the args ($fileName, $alias) -- they must
read/write from fileName, but now, they can also choose which method of
doing so based on the alias. Consider the definition of multiple
aliases, and a KWrap::new spewFrom argument:

 vlookup => [\&KWrap::lookup, 1, 0, [qw(actId)]],
 clookup => [\&KWrap::lookup, 1, 0, [qw(actId)]],

 spewFrom => sub {
   if ($_[1] eq "vlookup") {
     # print w/ vim readonly
   }
   elsif ($_[1] eq "clookup") {
     # print w/ cat
   }
   return 1;
 }

The user can now choose to use vlookup or clookup from the shell, which
is highly desirable freedom.

Note: Erroneously (in terms of documentation), this was partially
implemented in the last change, but will only be effectful hereafter.