Answered TODO about where bucket resolution fits
To change buckets -- to move to a different set of Acts where Acts have
different meanings -- from within one runKWrap; to create some arbitrary
number of KWrap objects, is incompatible with the way things are
currently set up. Command line arguments provide an easy alternative:

                     INPUT_SYMBOL     KWRAP_PATH
                        |   OUTPUT_SYMBOL  |
                        |       |          |
                       ####   ####   #############
 perl ... runKWrap.pl 'AB$ ' '  | ' 'ContextAyyBee'

Under such a configuration, an outer shell can start different subshells
for each context, with a matching unique abbrevation such as 'AB' so
that external ActId logging tools are given unique Ids when accompanied
by this prefix ...

This is not the place for rigorous specification, merely I offer an
explanation of why this idea has been dismissed at this time. But there
is some significance to it, so I don't doubt there will be proper
documention at some point in the future.