Command line arguments now captured correctly
Defining a $CONSOLE_OUTPUT_SYMBOL CLA (command line argument) now has an
effect.

As for why it didn't before, interestingly, the script

  my $A = "A1";
  our $B = "B1";

  sub whatAB {
    print "A = $A\n";
    print "B = $B\n";
  }

  sub mutateAB {
    our $A = "A2";
    our $B = "B2";
  }

  mutateAB;
  whatAB;

prints

  A = A1
  B = B2

So, variables declared as 'my' appear to be greedily captured into
closures statically, while declarations of 'our' are left as free. My
usage of this seems a bit exploitative, given I was heretofore
unawares. But, it was equally likely that I had chosen 'our' instead of
'my' initially, at which point things would have worked while I remained
none the wiser. Such is the miracle of abstraction.