Relaxed expectation for KWrap to handle undef
Calling, for example, $kw->edit(undef), would result in the creation of
an errorInfo value of the form "'$actId' is not an existing actId.".
This would try to perform string interpolation with an undef value,
raising a warning (but otherwise succeeding).

Currently, it is guaranteed by runKWrap that KWrap interface functions
are NEVER given any undef values, so these cases are temporarily
obsolete.